Add aria-hidden to svg chevrons
Certain AT announce these as "chevron-right" (uses the `<title>` present in
the SVG icon itself) under certain conditions (e.g. when they're part of a
button/role="button" element). This hides the SVG completely from AT,
suppressing this (as the meaning they're conveying is already covered by
the aria-expanded we dynamically added in a previous commit).